UI Changes, Saves Rating, Tags and Column Order
New icon for better visibility in the Windows Taskbar.
Tooltips in the settings window to explain certain elements.
Save and restore of the column order in the blog manager.
Properly save changes to the ratings, tags and notes (new free text
column) of the blogs after shutdown.
Code Cleanup (removed unnecessary usings and references).