Improved cpu performance.
- CPU profiling showed a high cycle count in the AddToDownloadList
method. Removing the duplicate check for the queue (producer consumer
collection) and catching the exception in the worst case of downloading
the same file twice within a short time period might be the better
option. For this to happen, two crawler tasks have to load the same file
from the queue, thus a duplicate post has to be made within a short
timeframe (unlikely).