Improved raw size handling
- Allows to specify which hosts to check for a _raw image version. The
failback now iterates through all sizes of the original found host until
a valid size was found that can be downloaded.
- The tumblr hosts can be set in the settings.json in the  TumblrHosts
variable. The first entry is check first, if there was no _raw image
found, the next entry is checked, until the list is exhausted. Then the
_1280 size of the original host is tried, the _500, _400 until the
request was successfully.