Improved image scaling, changed buffer size
- Scales the displayed image in the full screen preview Down if its
dimensions are larger than the screen size.
- Allows to set the buffer size for downloading binary files before
calling the WriteAsync() method on the file stream in the settings.json.
The previous default was 4KB, now it’s 2 MB. The max stack size for .NET
applications is 4MB (64 bit) or 1MB (32 bit). This might break 32bit
combability. It also increases memory usage depending on the number of
concurrent downloads but reduces disk fragmentation. Note: Setting the
file size via the SetLength() method kills the resuming as it
pre-allocates the whole file.
- Improves the handling of different image sizes in the crawler of
TumblrDownloader class. The returned json from the svc service seems to
vary the largest image size depending on its actual dimension instead of
always returning 1280.